Clearer instructions on database usage
In the second half of the exercise, the instructions imply that data should be moved from the `denormal_cars` database to the `normal_cars` database. This is incorrect. The `denormal_data` dataset should be inserted into the `normal_cars` database like the first half of the exercise and the data should be moved from the `car_models` table into the newly created normalized tables.